+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Move `get_bad_ocid_prefixes` from `libcoveocds.lib.common_checks` to `libcoveocds.common_checks`.
+- Move `libcoveocds.lib.additional_checks` to `libcoveocds.additional_checks`.
 
 ### Removed